Rework game board representation
Reworking the logic behind the game board to make it simpler to use.

Game not longer use Pygame sprites as the source of truth for virus/pill positions and collisions. Instead, a 16x8 gameBoard array represents the current state of the board. Pygame sprites update their positions based on the X,Y coords in the gameBoard array.